implemented bind maps to improve performance of binding

+++ src/binder/bind-number.cc
@@ -2,8 +2,8 @@
 // If no index is given, the next anonymous index is used.
 // If an error occurs, error is set to an appropriately descriptive string.
 
-void Binder::BindNumber(v8::Local<v8::Number> value, unsigned int index) {
+void Binder::BindNumber(v8::Local<v8::Number> value, int index) {
 	if (!index) {index = NextAnonIndex();}
-	int status = sqlite3_bind_double(handle, (int)index, value->Value());
+	int status = sqlite3_bind_double(handle, index, value->Value());
 	SetBindingError(status);
 }
